src/docker/api/app/requirements.txt
flask 
pyodbc
azure.storage.blob
azure.identity
python-decouple